add : 대기열 로직(실시간 알림은 미완)
