+++ src/main/java/com/docde/common/aop/CurrentCount.java
@@ -1,0 +1,11 @@
+package com.docde.common.aop;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Target(ElementType.METHOD)
+@Retention(RetentionPolicy.RUNTIME)
+public @interface CurrentCount {
+}
